Move ImageURLDialog to its own file
